feat: implement Onboarding Test Question 2 screen
Extract QuestionLayout component for reuse across question screens.
